+++ scout/utils/track_resources.py
@@ -8,6 +8,7 @@ LOG = logging.getLogger(__name__)
 REQUIRED_FIELDS = ["name", "type", "url"]
 TRACK_KEYS = ["name", "type", "format", "url", "indexURL"]
 URL_PATTERN = re.compile("https?://")
+REFERENCE_URL_KEYS = ["fastaURL", "indexURL", "cytobandURL", "aliasURL"]
 
 from typing import List
 
@@ -16,9 +17,12 @@ class AlignTrackHandler:
     """Class collecting external IGV tracks stored in the cloud"""
 
     def init_app(self, app):
-        self.tracks = self.set_custom_tracks(
-            app.config.get("CUSTOM_IGV_TRACKS") or app.config.get("CLOUD_IGV_TRACKS")
-        )
+        if app.config.get("CUSTOM_IGV_TRACKS") or app.config.get("CLOUD_IGV_TRACKS"):
+            self.tracks = self.set_custom_tracks(
+                app.config.get("CUSTOM_IGV_TRACKS") or app.config.get("CLOUD_IGV_TRACKS")
+            )
+        if app.config.get("CUSTOM_REFERENCE"):
+            self.custom_reference = self.set_custom_reference(app.config.get("CUSTOM_REFERENCE"))
 
     def track_template(self, track_info: dict) -> dict:
         """Provides the template for a VCF track object"""
@@ -44,6 +48,15 @@ class AlignTrackHandler:
             return abspath(path)
         else:
             raise FileNotFoundError(f"Could not verify path to IGV track:{path}")
+
+    def set_custom_reference(self, custom_references: dict) -> dict:
+        """Returns the complete path to a local igv track file"""
+
+        for custom_ref in custom_references.values():
+            for url_type in REFERENCE_URL_KEYS:
+                if url_type in custom_ref and not URL_PATTERN.search(custom_ref[url_type]):
+                    custom_ref[url_type] = self.set_local_track_path(custom_ref[url_type])
+        return custom_references
 
     def set_custom_tracks(self, track_dictionaries: List[dict]) -> dict:
         """Return a list of IGV tracks collected from the config settings